services/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas>=2.1.4
numpy>=1.26.3
yfinance>=0.2.33
xlsxwriter>=3.2.0
anthropic>=0.7.0
langchain-anthropic>=0.0.11
langchain>=0.1.9
langchain-community>=0.0.24
langchain-core>=0.1.27
aiohttp>=3.9.1
asyncio>=3.4.3
python-dotenv>=0.19.0
tweepy==4.14.0
praw==7.7.1
scikit-learn==1.3.0
transformers==4.31.0
torch==2.0.1
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose==3.3.0
boto3==1.28.1
pytest>=7.0.0
black==23.7.0
flake8==6.0.0
brotli>=1.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-multipart>=0.0.6
tiktoken>=0.5.1
httpx>=0.24.0
pytest-asyncio>=0.23.0